+++ ci/gpu/build.sh
@@ -74,13 +74,21 @@ logger "GoogleTest for libcudf..."
 cd $WORKSPACE/cpp/build
 GTEST_OUTPUT="xml:${WORKSPACE}/test-results/" make -j${PARALLEL_LEVEL} test
 
+logger "Python py.test for libcudf..."
+cd $WORKSPACE/cpp/build/python
+py.test --cache-clear --junitxml=${WORKSPACE}/junit-libgdf.xml -v
+
 # Temporarily install cupy for testing
 logger "pip install cupy"
 pip install cupy-cuda92
 
-logger "Python py.test for libcudf..."
-cd $WORKSPACE/cpp/build/python
-py.test --cache-clear --junitxml=${WORKSPACE}/junit-libgdf.xml -v
+# Temporarily install feather for testing
+logger "conda install feather-format"
+conda install -c conda-forge -y feather-format
+
+# Temporarily install tzdata otherwise pyarrow core dumps
+logger "apt-get update && apt-get install -y tzdata"
+apt-get update && apt-get install -y tzdata
 
 logger "Python py.test for cuDF..."
 cd $WORKSPACE/python